ET: diagram with sets not lists. Edges connect Plugs.
- A diagram is composed by a *set* of Nodes and Edges instead of a list.
That's important to make sure the algorithm doesn't depend on the order
or the elements.
- Removed the environment from the diagram. In the untyped lambda
calculus we don't need an environment neither in the language nor in the
diagram. For other languages we might need an environment.
- Edges connect Plugs instead of Nodes. Before, the order of the edges
was used to determine the corresponding hole in the node. Now, Nodes
contain a Plug (the place to connect an Edge on top of the node) and a
Hole contains a Plug. An Edge contains 2 Plugs.
- An Edge is undirected and that's reflected in its Eq and Ord.
- A Plug should be uniquely identifiable (i.e. contain unique ids) (WIP:
it works now but the type doesn't reflect that uniqueness. The diagram
could be created in an inconsistent state)
- Little improvement to unparse.
- Removed intermediate data structure of a rooted diagram. The correct
representation of a diagram algebraically is indeed the one with Maybe
root. That makes it possible to think of it as Magma with identity (an
algebraic Loop), where the identity is the empty diagram (with empty
sets of nodes and edges and a Nothing root). That also reduced the
amount of code.