TAPLMemoryDiagram for Lambda Ref done
The NM was addapted to reflect the "ala Racket" tracer that exposes the
information necessary to build the reference diagrams in TAPL.

As a result the core NM data structure changed and the f_NM now is step
and not low level memory manipulation operations (alloc, deref, etc).
(maybe there could be space for this lower level approach in the
future?)

The bisimulation and the tests were adapted accordingly.