Fix subtle issue in alligator render (height ratio alligator/egg)
How to calculate the with of each alligator family?

If an alligator with width `a` is protecting a list of families `fs` then the
sum of the widths of the families in `fs` must be `a`.
The first attempt was to equally distribute the width of the alligator among
the families in `fs`.
The problem is that the height of an egg may end up being bigger than the
height of an alligator.

To understand the solution to this consider the following example:

```
                a
      +--------------------+
      |                    |
      |                    |
      |         A          | a*a_p
      |                    |
      |                    |
      +-------------+------+
      |             |      |
b*b_p |      B      |   C  | c*c_p
      |             |      |
      +-------------+------+
             b          c
```

A and B are alligators and C is an egg.
A is protecting B and C.
The width of A is `a`, the width of B is `b` and the width of C is `c`.

The ratio between the width and the height of each sprite is fixed,
`a_p`, `b_p` and `c_p` respectively.

So if we have `n` alligators and `m` eggs in the list `fs` then our two constraints are:
- The sum of the widths of the families in `fs` must be `a`.
- The height of an egg must be equal to the height of an alligator.

These constraints are represented by the following equations (AsciiMath notation):

`{(n*b + m*c = a),
  (b*b_p = c*c_p):}`

Isolating `b` and `c` in the second equation:

`b=(c*c_p)/b_p` and `c=(b*b_p)/c_p`

Substituting `b` in the first equation:

`n*((c*c_p)/b_p)+m*c=a`

Isolating `c`:

`c=a/(n*((c_p)/b_p)+m)`

Substituting `c` in the first equation:

`n*b+m*((b*b_p)/c_p)=a`

Isolating `b`:

`b=a/(n+m*((b_p)/c_p))`

See good example with:
```
Alligator> :render (\a. (\d.a) a ((\a.a) c)) (\b.b)
```