LangToNM class, ExpTree diagram
The introduction of the LangToNM class allows for a more correct type-
level representation of sound notional machines as commutative diagrams,
mostly reflected in the Bisimulation, Injective, and Bijective classes
and related changes throughout the code. The change also leads to code
simplifications and more flexibility with improved static guarantees.